move to price_history file
